models/user: Move `User::rights()` fn into `controllers::helpers` module
