+++ src/models.rs
@@ -12,7 +12,6 @@ pub use self::follow::Follow;
 pub use self::keyword::{CrateKeyword, Keyword};
 pub use self::krate::{Crate, CrateName, NewCrate, RecentCrateDownloads};
 pub use self::owner::{CrateOwner, Owner, OwnerKind};
-pub use self::rights::Rights;
 pub use self::team::{NewTeam, Team};
 pub use self::token::ApiToken;
 pub use self::user::{NewUser, User};
@@ -32,7 +31,6 @@ mod follow;
 mod keyword;
 pub mod krate;
 mod owner;
-mod rights;
 pub mod team;
 pub mod token;
 pub mod user;
